pyproject.toml
[project]
name = "kolam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "matplotlib>=3.10.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
